refactor(lsp): extract workspace diagnostic modules
