+++ crates/tombi-lsp/src/lib.rs
@@ -13,6 +13,8 @@ mod references;
 mod remote_file;
 mod schema_resolver;
 mod semantic_tokens;
+mod workspace_config;
+mod workspace_diagnostic;
 
 pub mod handler {
     mod associate_schema;
@@ -79,7 +81,7 @@ pub mod handler {
     pub use shutdown::handle_shutdown;
     pub use update_config::handle_update_config;
     pub use update_schema::handle_update_schema;
-    pub use workspace_diagnostic::{handle_workspace_diagnostic, push_workspace_diagnostics};
+    pub use workspace_diagnostic::handle_workspace_diagnostic;
 }
 
 pub use backend::Backend;
